+++ lib/fmvvmnavigationservice.dart
@@ -16,9 +16,15 @@ class FmvvmNavigationService implements NavigationService {
 
     await Navigator.of(_viewContext)
         .pushNamed(routeName, arguments: _viewModel);
+
+    _viewModel.dispose();
   }
 
-  @override
+  /// Navigates to a new viewmodel of the type specified by the generic.
+  ///
+  /// The [parameter] is a value that will be passed to the new viewmodel's
+  /// init method. It is expexted that the ViewModel being navigated to will
+  /// return an instance of type R when it is popped off the stack.
   Future<R> navigateForResult<R extends Object, V extends ViewModel>(
       {Object parameter}) async {
     var _viewModel = createViewModel<V>(parameter);
@@ -28,10 +34,31 @@ class FmvvmNavigationService implements NavigationService {
     var returnValue = await Navigator.of(_viewContext)
         .pushNamed<R>(routeName, arguments: _viewModel);
 
+    _viewModel.dispose();
+
     return returnValue;
   }
 
-  @override
+  /// Navigates to a new view modeland removed the calling viewmodel from the stack.
+  ///
+  /// The [parameter] is a value that will be passed to the new viewmodel's
+  /// init method. The Future will be resolved with the ViewModel that is navigated to is
+  /// popped from the stack.
+  void navigateAndRemoveCurrent<V extends ViewModel>({Object parameter}) async {
+    var _viewModel = createViewModel<V>(parameter);
+
+    var routeName = Core._viewLocator.getViewFromViewModelType<V>();
+
+    await Navigator.of(_viewContext)
+        .popAndPushNamed(routeName, arguments: _viewModel);
+
+    _viewModel.dispose();
+  }
+
+  /// Any initiliazation needed to be done by the messenger service.
+  ///
+  /// This is called once when fmvvm is initialized. fmvvm uses a single
+  /// instance of the navigaiton service throughout the app.
   void initialize() {
     var messageService = Core.componentResolver.resolveType<MessageService>();
 
@@ -47,6 +74,7 @@ class FmvvmNavigationService implements NavigationService {
     Navigator.of(_viewContext).pop();
   }
 
+  /// Pops the current view / viewmodel off the stack and goes to the previous one.
   void navigateBackWithResult<R extends Object>([R parameter]) {
     Navigator.of(_viewContext).pop(parameter);
   }